feat: create event catalogue
Create event catalogues for a list of stations. These are used as the
input for the additional set of scripts that produce the inputs for
SplitRacer.

Event catalogues can be created from a csv file containing station
information, or (preferably) from a response file that can be read into
an `obspy.Inventory` object.